WIP web worker - abort

+++ web-client/src/components/SrAdvOptSidebar.vue
@@ -17,6 +17,7 @@
     import { useCurAtl06ReqSumStore } from '@/stores/curAtl06ReqSumStore';
     import { WorkerSummary } from '@/workers/workerUtils';
     import { WorkerMessage } from '@/workers/workerUtils';
+    import { WebWorkerCmd } from "@/workers/workerUtils";
 
     const reqParamsStore = useReqParamsStore();
     const sysConfigStore = useSysConfigStore();
@@ -37,15 +38,14 @@
     const iceSat2APIsItems = ref([{name:'atl06',value:'atl06'},{name:'atl06s',value:'atl06s'},{name:'atl03',value:'atl03'},{name:'atl08',value:'atl08'},{name:'atl24s',value:'atl24s'}]);
     const gediSelectedAPI = ref({name:'gedi01b',value:'gedi01b'});
     const gediAPIsItems = ref([{name:'gedi01b',value:'gedi01b'},{name:'gedi02a',value:'gedi02a'},{name:'gedi04a',value:'gedi04a'}]);
-    const isLoading = ref(false);
 
     let worker: Worker | null = null;
     type TimeoutHandle = ReturnType<typeof setTimeout>;
     let timeoutHandle: TimeoutHandle | null = null; // Handle for the timeout to clear it when necessary
 
     onMounted(() => {
-        console.log('SrAdvOptSidebar onMounted');
-
+        console.log('SrAdvOptSidebar onMounted totalTimeoutValue:',reqParamsStore.totalTimeoutValue);
+        mapStore.isAborting = false; 
     });
 
     watch(() => missionValue,(newValue,oldValue) => {
@@ -57,167 +57,8 @@
         }
     });
 
-    // async function runAtl06(req:SrRequestRecord){
-    //     console.log('runAtl06 with req:',req);
-    //     //console.log("runSlideRuleClicked typeof atl06p:",typeof atl06p);
-    //     //console.log("runSlideRuleClicked atl06p:", atl06p);
-    //     let recs:Elevation[] = [];
-    //     const callbacks = {
-    //         atl06rec: (result:any) => {
-    //             // if(cb_count.value === 0) {
-    //             //     console.log('first atl06p cb result["elevation"]:', result["elevation"]); // result["elevation"] is an array of Elevation');
-    //             // }
-    //             const currentRecs = result["elevation"];
-    //             const curFlatRecs = currentRecs.flat();
-    //             if(curFlatRecs.length === 0) {
-    //                 console.log('atl06p cb curFlatRecs.length === 0');
-    //                 return;
-    //             }
-    //             curReqSumStore.addNumRecs(curFlatRecs.length);
-    //             recs.push(curFlatRecs);
-    //             cb_count.value += 1;
-    //             updateExtremes(curFlatRecs);
-    //             const flatRecs = recs.flat();
-    //             //console.log(`flatRecs.length:${flatRecs.length} lastOne:`,flatRecs[flatRecs.length - 1]);
-    //             updateElevationLayer(flatRecs,true);
-    //             db.transaction('rw', db.elevations, async () => {
-    //                 try {
-    //                     // Adding req_id to each record in curFlatRecs
-    //                     const updatedFlatRecs: Elevation[] = curFlatRecs.map((rec: Elevation) => ({
-    //                         ...rec,
-    //                         req_id: requestsStore.currentReqId, 
-    //                     }));
-    //                     //console.log('flatRecs.length:', updatedFlatRecs.length, 'curFlatRecs:', updatedFlatRecs);
-    //                     await db.elevations.bulkAdd(updatedFlatRecs);
-    //                     //console.log('Bulk add successful');
-    //                 } catch (error) {
-    //                     console.error('Bulk add failed: ', error);
-    //                     requestsStore.setMsg('DB txn failed');                    
-    //                 }
-    //             }).catch((error) => {
-    //                 console.error('Transaction failed: ', error);
-    //                 requestsStore.setMsg('Transaction failed');
-    //                 toast.add({severity: 'error', summary: 'Transaction failed', detail: error.toString(), life: srToastStore.getLife() });
-    //             });            
-    //         },
-    //         exceptrec: (result:any) => {
-    //             console.log('atl06p cb exceptrec result:', result);
-    //             toast.add({severity: 'error',summary: 'Exception', detail: result['text'], life: srToastStore.getLife() });
-    //             requestsStore.setMsg(result['text']);
-    //             requestsStore.updateReq({req_id: req.req_id, status:'processing'});
-    //         },
-    //         eventrec: (result:any) => {
-    //             console.log('atl06p cb eventrec result:', result);
-    //             const this_detail = `Level:${result['level']}  ${result['attr']}`;
-    //             //toast.add({severity: 'info',summary: 'Progress', detail: this_detail, life: srToastStore.getLife() });
-    //             requestsStore.setMsg(this_detail)
-    //             requestsStore.updateReq({req_id: req.req_id, status:'processing'});
-    //         },
-    //     };
-    //     const map = mapStore.getMap() as OLMap ;
-    //     if (map){
-    //         console.log("atl06p cb_count:",cb_count.value)
-    //         isLoading.value = true; // controls spinning progress
-    //         if(req.req_id){       
-    //             requestsStore.reqIsLoading[req.req_id] = true; // for drawing control
-    //             console.log("atl06pParams:",req.parameters);
-    //             if(req.parameters){
-    //                 console.log('atl06pParams:',req.parameters);
-    //                 atl06p(req.parameters as Atl06pReqParams,callbacks)
-    //                 .then(
-    //                     () => { // result
-    //                         // Log the result to the console
-    //                         // Display a toast message indicating successful completion
-    //                         const flatRecs = recs.flat();
-    //                         console.log(`Final: flatRecs.length:${flatRecs.length} lastOne:`,flatRecs[flatRecs.length - 1]);
-    //                         if(flatRecs.length > 0) {
-    //                             updateElevationLayer(flatRecs,true);
-    //                             const status_details = `RunSlideRule completed successfully. recieved ${recs.flat().length} pnts`;
-    //                             toast.add({
-    //                                 severity: 'success', // Use 'success' severity for successful operations
-    //                                 summary: 'Success', // A short summary of the outcome
-    //                                 detail: status_details, // A more detailed message
-    //                                 life: 10000 // Adjust the duration as needed
-    //                             });
-    //                             requestsStore.updateReq({req_id: req.req_id,status: 'success' ,status_details: status_details});
-    //                         } else {
-    //                             const status_details = 'No data returned from SlideRule.';
-    //                             toast.add({
-    //                                 severity: 'error', // Use 'error' severity for error messages
-    //                                 summary: 'No Data returned', // A short summary of the error
-    //                                 detail: status_details, // A more detailed error message
-    //                             });
-    //                             console.log('Final: No more data returned from SlideRule.');
-    //                             requestsStore.updateReq({req_id: req.req_id, status: 'error', status_details: status_details});
-    //                         }
-
-    //                     },
-    //                     error => {
-    //                         // Log the error to the console
-    //                         console.log('runSlideRuleClicked Error = ', error);
-    //                         // Display a toast message indicating the error
-    //                         const status_details = `An error occurred while running SlideRule: ${error}`;
-    //                         toast.add({
-    //                             severity: 'error', // Use 'error' severity for error messages
-    //                             summary: 'Error', // A short summary of the error
-    //                             detail: status_details, // A more detailed error message
-    //                         });
-    //                         let emsg = '';
-    //                         if (navigator.onLine) {
-    //                             emsg =  'Network error: Possible DNS resolution issue or server down.';
-    //                         } else {
-    //                             emsg = 'Network error: your browser appears to be/have been offline.';
-    //                         }
-    //                         toast.add({
-    //                             severity: 'error',   
-    //                             summary: 'Error',   
-    //                             detail: emsg,      
-    //                         });
-    //                         requestsStore.updateReq({req_id: req.req_id,status: 'error', status_details: emsg});
-    //                     }
-    //                 ).catch((error => {
-    //                     // Log the error to the console
-    //                     console.error('runSlideRuleClicked Error = ', error);
-    //                     const status_details = `An error occurred while running SlideRule: ${error}`;
-    //                     requestsStore.updateReq({req_id: req.req_id,status: 'error', status_details: status_details});
-
-    //                     // Display a toast message indicating the error
-    //                     toast.add({
-    //                         severity: 'error', // Use 'error' severity for error messages
-    //                         summary: 'Error', // A short summary of the error
-    //                         detail: 'An error occurred while running SlideRule.', // A more detailed error message
-    //                     });
-    //                 })).finally(() => {
-    //                     const curAtl06ReqSumStore = useCurAtl06ReqSumStore();
-    //                     if(req.req_id){
-    //                         console.log('runAtl06 req_id:',req.req_id, ' updating stats');
-    //                         const extLatLon = {req_id: req.req_id, minLat: curAtl06ReqSumStore.get_lat_Min(), maxLat: curAtl06ReqSumStore.get_lat_Max(), minLon: curAtl06ReqSumStore.get_lon_Min(), maxLon: curAtl06ReqSumStore.get_lon_Max()};
-    //                         db.addOrUpdateExtLatLon(extLatLon);
-    //                         const extHMeanData = {req_id: req.req_id, minHMean: curAtl06ReqSumStore.get_h_mean_Min(), maxHMean: curAtl06ReqSumStore.get_h_mean_Max(), lowHMean: curAtl06ReqSumStore.get_h_mean_Low(), highHMean: curAtl06ReqSumStore.get_h_mean_High()};
-    //                         db.addOrUpdateHMeanStats(extHMeanData);
-    //                         isLoading.value = false; // for local button control
-    //                         requestsStore.reqIsLoading[req.req_id] = false; 
-    //                     } else {
-    //                         console.error('runAtl06 req_id was undefined?');
-    //                     }
-    //                     console.log(`cb_count:${cb_count.value}`)
-    //                 });
-    //             } else {
-    //                 console.error('runAtl06 req.parameters was undefined');
-    //             }
-    //         } else {
-    //             console.error('runAtl06 req.req_id was undefined');
-    //         }
-    //     }
-    // }
-
     const handleAtl06WorkerMsg = (event: MessageEvent) => {
         if(worker){
-            // Clear the timeout when any conclusive worker message is received
-            if (timeoutHandle) {
-                clearTimeout(timeoutHandle);
-                timeoutHandle = null;
-            }
             const workerMsg:WorkerMessage = event.data;
             console.log('handleAtl06WorkerMsg Worker event:',event);
             switch(workerMsg.status){
@@ -225,14 +66,18 @@
                     console.log('handleAtl06WorkerMsg success:',workerMsg.msg);
                     toast.add({severity: 'success',summary: 'Success', detail: workerMsg.msg, life: srToastStore.getLife() });
                     mapStore.isLoading = false;
-                    isLoading.value = false; // controls spinning progress
                     console.log('done... isLoading:',mapStore.isLoading);
-                    worker.terminate(); // Optionally terminate on success
-                    worker = null;
+                    cleanUpWorker(worker);
                     break;
                 case 'started':
                     console.log('handleAtl06WorkerMsg started');
                     toast.add({severity: 'info',summary: 'Started', detail: workerMsg.msg, life: srToastStore.getLife() });
+                    break;
+                case 'aborted':
+                    console.log('handleAtl06WorkerMsg aborted');
+                    toast.add({severity: 'warn',summary: 'Aborted', detail: workerMsg.msg, life: srToastStore.getLife() });
+                    requestsStore.setMsg('Job aborted');
+                    cleanUpWorker(worker);
                     break;
                 case 'server_msg':
                     console.log('handleAtl06WorkerMsg server_msg:',workerMsg.msg);
@@ -263,10 +108,7 @@
                         console.log('handleAtl06WorkerMsg error:',workerMsg.error);
                         toast.add({severity: 'error',summary: workerMsg.error?.type, detail: workerMsg.error?.message, life: srToastStore.getLife() });
                     }
-                    worker.terminate();
-                    worker = null;
-                    isLoading.value = false; // controls spinning progress
-                    mapStore.isLoading = false;
+                    cleanUpWorker(worker);
                     console.log('Error... isLoading:',mapStore.isLoading);
                     break;
                 default:
@@ -281,17 +123,30 @@
     function handleError(worker, error, errorMsg) {
         console.error('Error:', error);
         toast.add({ severity: 'error', summary: 'Error', detail: errorMsg, life: srToastStore.getLife() });
-        worker.terminate();
-        isLoading.value = false; // Ensure loading indicator is turned off
+        cleanUpWorker(worker);
+    }
+
+    function cleanUpWorker(worker){
+        if(worker){
+            worker.terminate();
+            worker = null;
+        }
+        if (timeoutHandle) {
+            clearTimeout(timeoutHandle);
+            timeoutHandle = null;
+        }
+        mapStore.isLoading = false; // controls spinning progress
+        mapStore.isAborting = false;
+        console.log('cleanUpWorker -- isLoading:',mapStore.isLoading);
     }
 
     function abortClicked() {
         if(worker){
-            worker.terminate();
-            worker = null;
-            isLoading.value = false; // controls spinning progress
-            console.log('abortClicked isLoading:',isLoading.value);
-            toast.add({severity: 'info',summary: 'Info', detail: 'Worker aborted', life: srToastStore.getLife() });
+            mapStore.isAborting = true; 
+            const cmd = {type:'abort',req_id:requestsStore.currentReqId} as WebWorkerCmd;
+            worker.postMessage(JSON.stringify(cmd));
+            console.log('abortClicked isLoading:',mapStore.isLoading);
+            requestsStore.setMsg('Abort Clicked');
         } else {
             console.error('abortClicked worker was undefined');
             toast.add({severity: 'error',summary: 'Error', detail: 'Worker was undefined', life: srToastStore.getLife() });
@@ -301,7 +156,7 @@
     async function runAtl06Worker(req:SrRequestRecord){
         try{
             if(req.req_id){
-                isLoading.value = true; // controls spinning progress
+                mapStore.isLoading = true; // controls spinning progress
                 requestsStore.currentReqId = req.req_id;
                 //requestsStore.reqIsLoading[req.req_id] = true; // for drawing control
                 const theUrl = new URL('@/workers/atl06ToDb', import.meta.url);
@@ -314,8 +169,10 @@
                         worker = null;
                     }
                 };
-                worker.postMessage(JSON.stringify(req));
+                const cmd = {type:'run',req_id:req.req_id, parameters:req.parameters} as WebWorkerCmd;
+                worker.postMessage(JSON.stringify(cmd));
                 const timeoutDuration = reqParamsStore.totalTimeoutValue*1000; // Convert to milliseconds
+                console.log('runAtl06Worker with timeoutDuration:',timeoutDuration, ' milliseconds');
                 timeoutHandle = setTimeout(() => {
                     if (worker) {
                         console.error('Timeout: Worker operation timed out in:',timeoutDuration);
@@ -337,7 +194,6 @@
     // Function that is called when the "Run SlideRule" button is clicked
     async function runSlideRuleClicked() {
         mapStore.isLoading = true;
-        isLoading.value = true;
         init(sysConfigStore.getSysConfig());
         console.log('runSlideRuleClicked isLoading:',mapStore.isLoading);
         let req = await requestsStore.createNewReq();
@@ -372,7 +228,6 @@
             }
         } else {
             mapStore.isLoading = false;
-            isLoading.value = false;
             console.error('runSlideRuleClicked req was undefined');
             toast.add({severity: 'error',summary: 'Error', detail: 'There was an error' });
         }
@@ -408,10 +263,10 @@
                 tooltipUrl="https://slideruleearth.io/web/rtd/api_reference/gedi.html#gedi"
             />
             <div class="button-spinner-container">
-                <Button label="Run SlideRule" @click="runSlideRuleClicked" :disabled="isLoading"></Button>
-                <Button label="Abort" @click="abortClicked" v-if:="isLoading"></Button>
-                <ProgressSpinner v-if="isLoading" animationDuration="1.25s" style="width: 3rem; height: 3rem"/>
-                <span v-if="isLoading">Loading... {{ curReqSumStore.getNumRecs() }}</span>
+                <Button label="Run SlideRule" @click="runSlideRuleClicked" :disabled="mapStore.isLoading"></Button>
+                <Button label="Abort" @click="abortClicked" v-if:="mapStore.isLoading" :disabled="mapStore.isAborting"></Button>
+                <ProgressSpinner v-if="mapStore.isLoading" animationDuration="1.25s" style="width: 3rem; height: 3rem"/>
+                <span v-if="mapStore.isLoading">Loading... {{ curReqSumStore.getNumRecs() }}</span>
             </div>
             <div class="sr-svr-msg-console">
                 <span class="sr-svr-msg">{{requestsStore.getConsoleMsg()}}</span>
